refactor: util의 getIdx를 viewModel안으로 옮김
